+++ share/objects/black-opal.lua
@@ -1,0 +1,15 @@
+local map = ...
+
+-- # Mystic Black Opal
+do
+	local unit = map.objects.h02K
+
+	-- ## Text
+	do
+		-- Hotkey
+		unit.uhot = {
+			type = 'string',
+			value = 'R'
+		}
+	end
+end